Moving functions to b_tensor_utils.py
